start migrating wallet tests into separate file (and mod for runner interface)
